requirements.txt
playwright
schedule
python-dotenv
pytest
pytest-mock
